Fix tradfri test
