+++ tests/components/tradfri/conftest.py
@@ -3,6 +3,8 @@ from unittest.mock import patch
 
 import pytest
 
+from tests.common import MockConfigEntry, mock_coro
+
 
 @pytest.fixture
 def mock_gateway_info():
@@ -11,3 +13,11 @@ def mock_gateway_info():
         "homeassistant.components.tradfri.config_flow.get_gateway_info"
     ) as mock_gateway:
         yield mock_gateway
+
+
+@pytest.fixture
+def mock_entry_setup():
+    """Mock entry setup."""
+    with patch("homeassistant.components.tradfri.async_setup_entry") as mock_setup:
+        mock_setup.return_value = mock_coro(True)
+        yield mock_setup
