+++ tests/components/tradfri/test_config_flow.py
@@ -18,14 +18,6 @@ def mock_auth():
         yield mock_auth
 
 
-@pytest.fixture
-def mock_entry_setup():
-    """Mock entry setup."""
-    with patch("homeassistant.components.tradfri.async_setup_entry") as mock_setup:
-        mock_setup.return_value = mock_coro(True)
-        yield mock_setup
-
-
 async def test_user_connection_successful(hass, mock_auth, mock_entry_setup):
     """Test a successful connection."""
     mock_auth.side_effect = lambda hass, host, code: mock_coro(
